requirements.txt
numpy==1.14.5
paho-mqtt==1.3.1
pkg-resources==0.0.0

